fix: Path Home

+++ nix/home/programs/editors/vscode.nix
@@ -3,15 +3,19 @@
 {
   home.file = {
     "Library/Application Support/Code/User/settings.json" = {
-      source = config.lib.file.mkOutOfStoreSymlink "${config.home.homeDirectory}/.dotfiles/nix/home/common/core/vscode/settings.json";
+      source = config.lib.file.mkOutOfStoreSymlink "${config.home.homeDirectory}/.dotfiles/nix/home/programs/editors/vscode/settings.json";
       onChange = ''
-        cp "$HOME/Library/Application Support/Code/User/settings.json" "${config.home.homeDirectory}/.dotfiles/nix/home/programs/editors/vscode/settings.json"
+        if [ -f "$HOME/Library/Application Support/Code/User/settings.json" ]; then
+          cp "$HOME/Library/Application Support/Code/User/settings.json" "${config.home.homeDirectory}/.dotfiles/nix/home/programs/editors/vscode/settings.json"
+        fi
       '';
     };
     "Library/Application Support/Code/User/keybindings.json" = {
       source = config.lib.file.mkOutOfStoreSymlink "${config.home.homeDirectory}/.dotfiles/nix/home/programs/editors/vscode/keybindings.json";
       onChange = ''
-        cp "$HOME/Library/Application Support/Code/User/keybindings.json" "${config.home.homeDirectory}/.dotfiles/nix/home/programs/editors/vscode/keybindings.json"
+        if [ -f "$HOME/Library/Application Support/Code/User/keybindings.json" ]; then
+          cp "$HOME/Library/Application Support/Code/User/keybindings.json" "${config.home.homeDirectory}/.dotfiles/nix/home/programs/editors/vscode/keybindings.json"
+        fi
       '';
     };
   };
